requirements.txt
Django==1.11.7
ipython==6.2.1
django-extensions==1.9.7
Pillow==4.3.0
djangorestframework==3.7.3
